pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "erp-platform"
version = "1.0.0"
description = "AI-Powered Modular ERP Platform"
requires-python = ">=3.10"
dependencies = [
    "Jinja2>=3.0",
    "Werkzeug>=2.0",
    "psycopg2-binary>=2.9",
]
